requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.0.1
certifi==2024.12.14
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.109.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.26.0
idna==3.10
iniconfig==2.0.0
numpy==2.2.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.10.1
pytest==7.4.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
pytz==2024.2
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.27.0
